requirements.txt
Flask==3.0.0
pybit==5.6.2
python-dotenv==1.0.0
gunicorn==21.2.0
google-cloud-datastore==2.19.0
rich==13.7.0
